requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
asyncpg
twilio
google-genai
boto3
PyJWT
passlib[bcrypt]
email-validator
python-multipart
